requirements.txt
asgiref==3.9.1
bcrypt==4.3.0
dj-database-url==3.0.1
Django==5.2.4
django-cors-headers==4.7.0
djangorestframework==3.16.0
gevent==25.5.1
greenlet==3.2.3
gunicorn==23.0.0
packaging==25.0
psycopg2-binary==2.9.10
python-dotenv==1.1.1
sqlparse==0.5.3
typing_extensions==4.14.1
whitenoise==6.9.0
zope.event==5.1
zope.interface==7.2
